requirements.txt
pyautogui
pywinauto
pillow
psutil
mouse
pydirectinput
sneakysnek
redis
pydantic